# server/swagger.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
  description: Production-ready REST API for managing authenticated users, projects, tasks, and dashboard analytics.
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    get:
      summary: List all projects accessible to the authenticated user
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Projects retrieved successfully
    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Project created successfully
...
  /api/tasks:
    get:
      summary: List accessible tasks with filtering, search, sorting, and pagination
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [To Do, In Progress, Done]
        - in: query
          name: priority
          schema:
            type: string
            enum: [Low, Medium, High]
        - in: query
          name: assignee
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Tasks retrieved successfully
    post:
      summary: Create a new task inside a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, projectId]
              properties:
                title:
                  type: string
                description:
                  type: string
                projectId:
                  type: string
                status:
                  type: string
                  enum: [To Do, In Progress, Done]
                priority:
                  type: string
                  enum: [Low, Medium, High]
                dueDate:
                  type: string
                  format: date
                assignee:
                  type: string
      responses:
        '201':
          description: Task created successfully
...